Move Diagnostics classes to impl and mark @internal
Each now sits in the `impl` directory alongside the other internals it reports on, and is dropped
from its package index - matching how `StoreValidator` and `RecordSet` are handled. The classes
remain reachable as the type of the public `diagnostics` properties.
